+++ lib/modular_router.dart
@@ -1,5 +1,5 @@
-library module_router;
-
 export 'src/model/module.dart';
 export 'src/model/route.dart';
-export 'src/router.dart';
+export 'src/model/injector.dart';
+export 'src/modular_router.dart';
+export 'src/navigator_extention.dart';