+++ lib/src/model/injector.dart
@@ -1,0 +1,7 @@
+class Injector<T> {
+  final Function constructor;
+
+  Injector(this.constructor);
+
+  Type get injectedType => T;
+}