+++ lib/src/model/route.dart
@@ -1,19 +1,26 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
+import 'package:stateful_widget_binder/stateful_widget_binder.dart';
 
-class ModuleRoute<T> {
+class ModuleRoute<T extends StatefulWidgetBinder> {
   final String path;
+  final Function viewBuilder;
+  final Function controllerBuilder;
   final bool allowAnonymous;
   final bool isFullscreenDialog;
-  final Widget Function() builder;
-  final bool keepAlive;
-
-  Type get type => T;
+  final PageTransitionBuilder? customPageTransition;
 
   ModuleRoute({
     required this.path,
     this.allowAnonymous = false,
     this.isFullscreenDialog = false,
-    required this.builder,
-    this.keepAlive = false,
+    required this.viewBuilder,
+    this.controllerBuilder = _defaultControllerBuilder,
+    this.customPageTransition,
   });
+
+  static dynamic _defaultControllerBuilder() => null;
+
+  Type get viewRuntimeType => T;
 }
+
+typedef PageTransitionBuilder = PageRoute<T> Function<T>({RouteSettings? settings, required Widget view});