+++ lib/src/model/routing/routing.dart
@@ -1,0 +1,11 @@
+class Routing {
+  final Type view;
+  final Type module;
+  final String path;
+
+  Routing({
+    required this.view,
+    required this.module,
+    required this.path,
+  });
+}