+++ lib/src/views/unauthorized_route.dart
@@ -1,8 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:stateful_widget_binder/stateful_widget_binder.dart';
 
-class UnauthorizedRoute extends StatelessWidget {
-  const UnauthorizedRoute();
+class UnauthorizedRoute extends StatefulWidgetBinder {
+  const UnauthorizedRoute({required super.controller});
 
+  @override
+  State<UnauthorizedRoute> createState() => _UnauthorizedRouteState();
+}
+
+class _UnauthorizedRouteState extends State<UnauthorizedRoute> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -12,9 +18,9 @@ class UnauthorizedRoute extends StatelessWidget {
           'Unauthorized access',
           textAlign: TextAlign.center,
           style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                fontWeight: FontWeight.bold,
-                fontSize: 36.0,
-              ),
+            fontWeight: FontWeight.bold,
+            fontSize: 36.0,
+          ),
         ),
       ),
     );